Start using size_t to use the maximum size of an int object - to be a…

+++ matrix.c
@@ -2,7 +2,7 @@
 
 Matrix_t matrix_load_from_file(FILE *file)
 {
-  long i, j;
+  size_t i, j;
   double rows, cols, value;
   fscanf(file, "%lf %lf", &rows, &cols);
   Matrix_t matrix = matrix_create(rows, cols);
@@ -13,17 +13,17 @@ Matrix_t matrix_load_from_file(FILE *file)
   return matrix;
 }
 
-Matrix_t matrix_create(long rows, long cols)
+Matrix_t matrix_create(size_t rows, size_t cols)
 {
   Matrix_t A;
   A.rows = rows;
   A.cols = cols;
   A.items = malloc(cols * rows * sizeof(double));
 
-  for (long i = 0; i < rows; i++)
+  for (size_t i = 0; i < rows; i++)
   {
     *(A.items+i) = malloc(cols * sizeof(double));
-    for (long j = 0; j < cols; j++)
+    for (size_t j = 0; j < cols; j++)
       A.items[i][j] = 0.0;
   }
 
@@ -33,13 +33,13 @@ Matrix_t matrix_create(long rows, long cols)
 Matrix_t matrix_transpose(Matrix_t a)
 {
   Matrix_t b = matrix_create(a.cols, a.rows);
-  for (long j = 0; j < a.cols; j++)
-    for (long i = 0; i < a.rows; i++)
+  for (size_t j = 0; j < a.cols; j++)
+    for (size_t i = 0; i < a.rows; i++)
       matrix_set(b, j, i, a.items[i][j]);
   return b;
 }
 
-void matrix_set(Matrix_t a, long row, long col, double value)
+void matrix_set(Matrix_t a, size_t row, size_t col, double value)
 {
   a.items[row][col] = value;
 }
@@ -48,10 +48,10 @@ Matrix_t matrix_mult(Matrix_t a, Matrix_t b)
 {
   Matrix_t result = matrix_create(a.rows, b.cols);
 
-  for (int i=0; i<a.rows; i++) {
-    for (int j=0; j<b.cols; j++) {
+  for (size_t i=0; i<a.rows; i++) {
+    for (size_t j=0; j<b.cols; j++) {
       double sum = 0;
-      for (int k=0; k<b.rows; k++) {
+      for (size_t k=0; k<b.rows; k++) {
         sum += a.items[i][k] * b.items[k][j];
       }
       result.items[i][j] = sum;
@@ -64,10 +64,10 @@ Matrix_t matrix_mult_t(Matrix_t a, Matrix_t b)
 {
   Matrix_t result = matrix_create(a.rows, b.rows);
 
-  for (int i=0; i<a.rows; i++) {
-    for (int j=0; j<b.rows; j++) {
+  for (size_t i=0; i<a.rows; i++) {
+    for (size_t j=0; j<b.rows; j++) {
       double sum = 0;
-      for (int k=0; k<b.cols; k++) {
+      for (size_t k=0; k<b.cols; k++) {
         sum += a.items[i][k] * b.items[j][k];
       }
       result.items[i][j] = sum;
@@ -83,8 +83,8 @@ bool matrix_mult_valid(Matrix_t a, Matrix_t b)
 
 void matrix_print(Matrix_t a)
 {
-  for (int i=0; i<a.rows; i++) {
-    for (int j=0; j<a.cols; j++) {
+  for (size_t i=0; i<a.rows; i++) {
+    for (size_t j=0; j<a.cols; j++) {
       printf("%03lf ", a.items[i][j]);
     }
     printf("\n");
@@ -93,10 +93,8 @@ void matrix_print(Matrix_t a)
 
 void matrix_print_to_file(FILE *output, Matrix_t a)
 {
-
   fprintf(output, "%ld %ld\n", a.rows, a.cols);
-  for (long i = 0; i < a.rows; i++)
-    for (long j = 0; j < a.cols; j++)
+  for (size_t i = 0; i < a.rows; i++)
+    for (size_t j = 0; j < a.cols; j++)
       fprintf(output, "%ld %ld %lf\n", i, j, a.items[i][j]);
-
 }
